+++ microwler/core/crawler.py
@@ -12,8 +12,8 @@ import completely
 from diskcache import Index
 from lxml import html as DOMParser
 
-from microwler.scrape import Page
-from microwler.settings import Settings
+from microwler.core.scrape import Page
+from microwler.core.settings import Settings
 from microwler import utils
 
 LOG = logging.getLogger(__name__)
